ai-services/notes_service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
reportlab
cryptography
